requirements.txt
Django==1.3.1
Pillow==1.7.6
certifi==0.0.8
chardet==1.0.1
distribute==0.6.24
django-imagekit==1.1
docutils==0.8.1
requests==0.10.6
wsgiref==0.1.2
